[Client] Rewiering the different controls in the client.
Almost done with this client iteration.

+++ Dev/Clients/QtClient/QtWindowClient/SWIUtils.cpp
@@ -3,7 +3,7 @@
 
 #include "SWIUtils.h"
 
-QString SWIUtils::FormatDownQuantity(int quantity)
+QString SWIUtils::FormatDownQuantity(int64_t quantity)
 {
     QString result;
